third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/insert-node-expected.txt

Tests that elements panel updates dom tree structure upon node insertion.


Running: testDumpInitial
========= Original ========
- <div id="container">
      <div id="child1"></div>
      <div id="child2"></div>
      <div id="child3"></div>
  </div>

Running: testInsertBefore
===== Inserted before =====
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child2"></div>
      <div id="child3"></div>
  </div>

Running: testInsertMiddle
===== Inserted middle =====
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child-middle"></div>
      <div id="child2"></div>
      <div id="child3"></div>
  </div>

Running: testAppend
======== Appended =========
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child-middle"></div>
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child-after"></div>
  </div>

Running: testAppendWithText
======== Appended with text=========
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child-middle"></div>
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child-after"></div>
      <div id="child-with-text" style="display: none;">Text</div>
  </div>
Success: child text is bound

Running: testInsertFirstTextNode
======== Inserted first text node =========
- <div id="container">
      <div id="child-before"></div>
      <div id="child1"></div>
      <div id="child-middle"></div>
      <div id="child2"></div>
      <div id="child3">First text</div>
      <div id="child-after"></div>
      <div id="child-with-text" style="display: none;">Text</div>
  </div>
Success: child text is bound
